requirements.txt
# requirements.txt
yfinance
pandas
numpy
matplotlib
seaborn
requests
pyyaml
scikit-learn
xgboost
tensorflow
torch
transformers
langchain
chromadb
streamlit
flask
newsapi-python
